+++ bikeability/artifact.py
@@ -10,11 +10,7 @@ from climatoology.base.artifact import (
 from climatoology.base.computation import ComputationResources
 
 from bikeability.input import PathRating
-from bikeability.utils import (
-    pathratings_legend_fix,
-    get_qualitative_color,
-    PathCategory,
-)
+from bikeability.utils import PathCategory, pathratings_legend_fix, get_qualitative_color
 
 
 def build_paths_artifact(
@@ -29,17 +25,21 @@ def build_paths_artifact(
     paths_polygon = paths_polygon.clip(clip_aoi, keep_geom_type=True)
     paths_all = pd.concat([paths_line, paths_polygon], ignore_index=True)
 
-    paths_all['color'] = paths_all.category.apply(get_qualitative_color, cmap_name=cmap_name)
+    paths_without_restriction = paths_all[paths_all.category.isin(PathCategory.get_visible())]
+
+    paths_without_restriction['color'] = paths_without_restriction.category.apply(
+        get_qualitative_color, cmap_name=cmap_name
+    )
     return create_geojson_artifact(
-        features=paths_all.geometry,
+        features=paths_without_restriction.geometry,
         layer_name='Cycling infrastructure path categories',
         caption=Path('resources/info/path_categories/caption.md').read_text(),
         description=Path('resources/info/path_categories/description.md').read_text(),
-        label=paths_all.category.apply(lambda r: r.name).to_list(),
-        color=paths_all.color.to_list(),
+        label=paths_without_restriction.category.apply(lambda r: r.name).to_list(),
+        color=paths_without_restriction.color.to_list(),
         legend_data={
             pathratings_legend_fix.get(category.value, category.value): get_qualitative_color(category, cmap_name)
-            for category in PathCategory
+            for category in PathCategory.get_visible()
         },
         resources=resources,
         filename='cycling_infrastructure_path_categories',